Extract common resolver API types.
